Add C bindings for prover
Add arbitrator_(load_machine, free_machine, clone_machine, step, hash,
gen_proof, free_proof). Machine is returned as a pointer to an
incomplete type, to be used as a handle only for passing to the other
functions, and it must be freed after use. The hex-stringified proof
must also be freed after use.

Headers are generated with cbindgen.

Tested with a simple C program, see notes on PR.