Add Broadcaster for Nitro
Implement a message broadcaster for Nitro that sends
arbstate.MessageWithMetadata messages using the WSBroadcasterServer.
This takes a new dependency on the arbitrum arb-util package.

The new Broadcaster caches messages to be sent to new clients in the
SequenceNumberCatchupBuffer, and supports clearing confirmed messages
by sequence number.

Added a test for the SequenceNumberCatchupBuffer and an example of the
serialization.