Add timeout parameter to DAS Store interface
When multiple DASes are used together with the aggregator, they all
need to have the same timeout, since the timeout is one of the data
fields that get signed. If different DASes have a different timeout,
their signatures cannot be aggregated together since they have signed
different things.

Added a timeout field to the protocol buffers spec, keeping it as a
simple uint64. There is an optional keyword available for primative
fields but the guidance from the protobuf developers is to just use
a plain integer. The default zero value is treated as unassigned and
callees should assign a timeout if this is seen.
See https://cloud.google.com/apis/design/design_patterns.md#optional_primitive_fields