Support complex DAS configurations
The DAS used by the node and daserver executables can now be configured
to use multiple types of persistent storage, multiple types of caching,
the REST client aggregator (alone or as a fallback), the RPC aggregator,
and to set up the whole stack allowing Store calls or not
(AllowStoreOrigination). The configuration is specified in a flat way
for ease of use, with the components instantiated according to a
predefined topology that we think will match the most common deployments
described in the comment inside SetUpDataAvailability in node.go.

Some notable changes made during the refactoring were:
1) arbstate.SimpleDASReader was renamed to DataAvailabilityReader and
   arbstate.DataAvailabilityServiceReader was deleted entirely, being
replaced by DataAvailabilityReader. KeysetFromHash and
CurrentKeysetBytes were the two methods in DataAvailabilityServiceReader
beyond GetByHash from SimpleDASReader. CurrentKeysetBytes was unused
beyond exposing it as an RPC method that was not used anywhere in the
system. KeysetFromHash had various implementations in different DAS
implementations, but it's unclear if getting the keyset hash of a leaf
storage DAS would actually be useful; the behavior is now simplified to
run the ChainFetchDAS L1 lookup on a failed GetByHash on the DAS stack.
2) local_disk_das.go was deleted and the main das class from it is now
   called SignAfterStoreDAS to indicate that it's an abstract DAS that
signs after storing to a StorageService.
3) LocalDiskStorageService is now called LocalFileStorageService and it
   got its own file.
4) A read limiting wrapper for the DataAvailabilityService interface was
   created to go along with the wrapper for StorageService, to allow
uniformity of types in the setup functions. The entire DAS stack is
wrapped in the read limiting wrapper if AllowStoreOrigination. We
continue to cast it back to DataAvailabilityReader where possible, and
if it is used as a DataAvailabilityService accidentally it is set to
panic so we would know immediately if we have improperly used it.