Options for pinning batch data fetched from IPFS
This enables Nitro nodes to also pin the batch data in IPFS,
with a configurable percent of batches to be pinned if the operator is
concerned about storage. Back of the envelope reasoning suggests
that the additional storage requirements are not more than the existing
storage requirements for a Nitro archive node, since an archive node
also stores all Nitro blocks.