coordinator: add a comment about sequencer pause race-condition
There is no race because sequencer is already either paused or
forwarding.
There could theoretically be some weird edge case where we somehow
lost and regained the lockdown very quickly. In this case - it
doesn't really matter if we scheduled the tx in parallel before or
after the delayed tx as there isn't really a need to schedule the
delayed tx first.