Remove DataAvailabilityService interface
Removing the DataAvailabilityService interface allows us to clean up a
lot of code. DataAvailabilityServiceReaders or Writers are now used
instead of the DataAvailabilityService interface.

Cache setup was moved to its own function and to apply directly to
StorageServices, allowing us to remove the CacheStorageToDASAdapter.

ChainFetchDAS was removed since only ChainFetchReader is needed.

Simplify CreateDAReaderWriterForStorage by requiring some kind of
storage to be configured.