Add relay feed connection limiting
This change adds four new configuration options to the relay executable.
--node.feed.output.connection-limits.enable
--node.feed.output.connection-limits.per-ip-limit
--node.feed.output.connection-limits.per-ipv6-cidr-48-limit
--node.feed.output.connection-limits.per-ipv6-cidr-64-limit
per-ip-limit controls the number of times a single IP may be connected
to this relay. The ipv6-cidr options control how many times any CIDR
block either /48 or /64 can be connected to this relay.
All connection-limit configuration is hot reloadable.

Websocket connection requests will almost always be rejected at the
handshake request to upgrade to websockets from the client by returning
a 429 response. In the event of a race where two or more handshakes are
completed when one of them should be rejected, whichever one gets
registered by the ClientManager last will be rejected by closing the
connection.

Private and loopback IP ranges are not blocked, and a Warning message is
printed for them since it indicates a misconfiguration if they are seen.

This change also fixes client IP detection for Cloudflare, we had not
canonicalized the header we were looking for.