Update to build against geth merge-v1.11.4 branch
Most of the changes were related to upstream's core.Message change.
core.Message has been changed from interface to struct and types.Message
has been removed. We used the core.Message interface in a number of
places and it needed to be changed to a pointer to the core.Message
struct to maintain the same semantics.