Reject RPC requests if memory limit is exceeded
If the --node.resource-mgmt.mem-limit-percent option is used then if
system memory usage exceds that limit, reject new RPC requests with a
HTTP 429 error.

If the option is used, then Nitro attempts on startup to discover what
method is available to check the system memory usage and limit.
Currently Cgroups V1 is the only supported method, and if it is not
detected then an error is logged once and the limit will not be
enforced.