Plumb batch limit options to geth
Upstream implemented limits on the number of requests in a batch and the
maximum batch response size. We had already implemented the batch
response size limit in our geth fork, but have rolled it back in favor
of upstream. This commit adds a nitro option for the batch request limit
and takes the existing size limit option and plumbs them through to
geth's node.Config.

The flags in Nitro are:
--rpc.max-batch-response-size
--rpc.batch-request-limit