Add loggers for ephemeral errors
These loggers can be set up to log to one underlying logging function if
if the time or count threshold since the first time it was logged with
has not yet been reached, and another log function for if the threshold
has been reached. An example use case would be to log a message at Debug
level, until the same message has been seen 5 times, or for 1 minute,
after which it would be logged at Error level until reset.