Use SimulatedBeacon for test L1s
v1.12.1 geth changes the default ChainConfig returned with
DeveloperGenesisBlock from AllCliqueProtocolChanges to
AllDevChainProtocolChanges. Since the dev chain config is not set to
clique mode and is set as post-TTD with TTD=0, there needs to be a
non-ethhash based consensus engine running in order for blocks to be
produced. See the upstream geth change here:
https://github.com/ethereum/go-ethereum/pull/27327

This commit copies the initailization code for how geth is started up
in dev mode with the simulated beacon node to our test L1 setup code.
It also enables FullSync mode on the L1, since otherwise
SimluatedBeacon.sealBlock would fail when calling the
ConsensusAPI.ForkChoiceUpdatedV2 as it would still be syncing if it was
in SnapSync mode. UseMergeFinality was also disabled on the
TestDelayedSequencerConfig since the SimulatedBeacon seems to have
different finalization behavior.