Batch txs rather than on-demand L1 blocks for tests
SimulatedBeacon creates a block per tx when running in on-demand mode,
and then increments the block timestamps by one if they would've been
for the same time, which can break assumptions in a lot of our tests,
for example conditional tx tests that set time range restrictions on the
txs. By setting SimulatedBeacon's delay to 1s, this allows it to batch
multiple txs into a block and avoid this problem.