Parallelize initial contract deployment
This is to make the tests go faster if on demand block mode is not used.
It has at least two problesm though.

Unfortunately due to the block gas limit, creating all the initial
rollup contracts still takes multiple blocks and causes a multi-second
delay in the tests. It can also end up pricing transactions incorrectly
by submitting them all in parallel.

Pushing to a branch for now, may revisit.