Avoid freq switch of non-4844 to 4844 batch post
Logic to prevent switching from non-4844 batches to 4844 batches too often,
so that blocks can be filled efficiently. The geth txpool rejects txs for
accounts that already have the other type of txs in the pool with
"address already reserved". This logic makes sure that, if there is a backlog,
that enough non-4844 batches have been posted to fill a block before switching.