Avoid deadlocking sender account w/ diff type txs
geth's current mempool implementation doesn't allow both blob and
non-blob txs at the same time for the same account. It has has separate
pools for each tx type, with a layer aggregating them and rejecting new
txs for an account if there are already txs for that account in the pool
of the other type.

This poses a hazard where Nitro could send batch txs of one type that
are evicted before being included in a block (eg due to GasFeeCap or
BlobFeeCap being too low), then for a new batch Nitro switches batch type
and sends a tx with higher nonce which is not rejected because the
parent mempool is currently empty. Then the presence of that tx would
prevent the earlier txs from being able to be re-sent. A similar situation
could arise where the mempool is gapped due to eventual consistency
between p2p nodes.

This commit makes it so a tx of a different type to the previous will
only be sent by the DataPoster if the previous tx has been included in
a block that has some reorg resistance (head-1). The BatchPoster will
continue making new batch txs and requesting the DataPoster to send
them, but if they fail this check they will just be queued and sent
in the DataPoster.Start loop.