Introduce mark-and-sweep dirty branch tracking.
This change introduces internal mutability to the merkle tree
implementation in the form of a dirty_indices HashSet<usize> which
keeps track of the indices of the parent nodes for leaves
which have changed since the last time the hashes for all of
the non-leaf nodes in the tree were calculated.

The `set` method was changed to only modify the leaf's value and
add the parent node to the set of dirty indices.

The hashes of all the non-leaf nodes which are dirty and all their
ancestors up to the root node are recalcualted just before returning
the value of the root.

This should allow us to reuse previously allocated merkle trees
rather than having to throw them out and recaculate the hashes
of all the nodes every time a leaf's value is changed.