Maybe 10x performance improvement.
This change has the internal mutability implemented with the
`Arc<Mutex<T>>` pattern as well as changing the `memory.rs` to no
longer pre-allocate the empty hashes for all of the memory it wants to
consume. There is still some inefficency in that the
`Merkle::resize()` method is still allocating a bunch of actual leaf
entries when it doesn't need to. We should be allocating those when
needed by the `Merkle::set` calls.

It's also possible that this code doesn't actually work when tied into
a real machine. Testing that next.