Update the zerohashes to use the ones from memory.rs.
Before this change, the MerkleType::MEMORY was loosely coupled to
the Memory class. But, the Memory::merkleize() method was doing a lot
of work to pass in a long vector of empty leaf hashes, and overriding
the default empty node's hash.

This change makes it the responsibility of the zerohashes constants to
know what the correct empty node hashes are for each layer in a MEMORY
merkle tree. But, that just happens to be the same value that the
Memory::merkleize method would have used if it were allowed to specify
it. This way, the code doesn't break backward compatibility with
anywhere the old values were stored.