Change log-level cli opt to take a string
The log-level cli opt on all executables built by this project now takes
a string instead of an integer, and the strings are the typical
CRIT, ERROR, WARN, INFO, DEBUG, TRACE. Additionally it accepts the old
geth numeric log levels as an undocumented feature to avoid the need for
anyone already using them to migrate their config.

The introduction of slog as geth's logger changed the numeric constants
used for the typical levels, so we make this change to avoid requiring any
config changes and also to give clearer options going forward.