Add more counters.
We need to be able to see how frequently the memory is being hashed
and from where.

The problem is that with always_merkleize off, there are thousands of
fewer calls to `Merkle::root()` and this is to help us figure out
where the calls are coming from.

Spoiler alert: It's from the flush_module macro.