Remove flush_module macro.
The high-level benefit of this change is a 16.5% reduction in block
processing time from ~17.6s to ~14.7s.

The only job of that macro was to recalculate the hash of the module
and update the module merkle tree's corresponding leaf with the new
hash.

Instead, the waits until the hash for the module is about to be
calculated and calls `Merkle::set()` for each module in the modules
vector.

This eliminates thousands of calls to the possibly expensive
`Merkle::root()` call on the `Memory` of each module.

This pattern of delaying when to actually update the leaves of merkle
trees until they will be used has a lot of potential for additional
performance boosts.