Make clone create completely independent layers structs.
The default behavior of an Arc is to allow sharing between different
threads and different instances. So, the derived behavior for cloning
was to put the same references on the clone as on the orignal.

But, we actually want separate clones of our merkle trees to be
independent. That is, changing the data on the clone, should have no
affect on the original (and vice versa.)

The test added in this commit failed before the change and passes
afer.