init: add option to download latest db snapshot
If --init.latest is set, get the latest date from
<mirror>/<chain-name>/latest-<snapshot-kind>.txt.

(The snapshot-kind can be "archive" or "pruned".)

The latest file contains a date in the format <year>/<week>.

Then, set the download URL to
<mirror>/<chain-name>/<year>/<week>/<snapshot-kind>.tar.

Finally, the code tries to download this snapshot as a whole file. If it
is not found (404) it tries to download it in parts.