Add a utility for getting a free TCP port.
This is especially useful in test code when you want to be able to
bring up local servers in your tests without conflicting with a test
that might be running in parallel.

Currently, the tests tagged challengetest fail on CI when they are run
in parallel with each other because only one test is able to bring up
a Signer on port 1234. The others silently fail to bring up a signer,
and the signer brought up by the first one may not still be around
when subsequently started parallel tests need one.