l1SyncService should sync full not chunked payload
Previously the l1SyncService was fetching the payload for each batch
and, instead of storing the batch itself, chunking the batch using the
dastree scheme and storing the chunks using the preimageRecorder with
RecoverPayloadFromDasBatch. But DAS batches are requested by hash of the
whole payload, and not by chunks, so the data being synced was useless.

This change fixes the l1SyncService to store the whole batch payload and
also to fix the logic around the sync-to-storage.acheck-already-exists option
so that it works correctly.

Any daservers that are currently running with sync-to-storage enabled
will restart syncing from sync-to-storage.eager-lower-bound-block
since they have previously synced incorrectly.