Switch to ref counted expiry using hard links
We can have batches with identical data, so the trieLayout needs to be
able to handle being requested to store the same batch multiple times
with different expiry times.

The way it works now is that the leaf files in by-expiry-timestamp are
hard links to the batch files in by-data-hash. The link count is used to
know when there are no more links in by-expiry-timestamp pointing to the
batch file and that it can be deleted.