Deprecate local-db-storage, add migration opt
The local DB storage backend, based on badgerdb, is being deprecated due
to its unfavorable memory performance when compacting.
local-file-storage is now the recommended local storage instead.
To migrate to it, use --data-availability.migrate-local-db-to-file-storage
and enable the local-file-storage service.

Migration must be enabled manually using this flag since the
operator may need to set up a new storage directory and backing volume.

In-process migration is only provided from local-db-storage to
local-file-storage since it is the only other local storage backend
we current provide. To migrate to other backends, start a new daserver
instance and use the rest-aggregator.sync-to-storage options.
In the case where both S3 and DB backends were being used, it is recommended
to migrate to the file backend and retain S3 instead of just dropping the DB
backend configuration, otherwise there would no longer be a local store.

ERROR and WARN messages on startup will inform operators that they
should migrate (but can continue anyway), if migration failed, or if
migration had previously succeeded and config/storage cleanup is
recommended. These messages will be made fatal in the future before the
local-db-storage backend code is removed entirely for the full
deprecation.