Metric for any das.Aggregator Store errors
If there are errors sending to any backends in das.Aggregator.Store then
the new metric arb_das_rpc_aggregator_store_error_gauge will be 1 until
there is a call to Store with no errors. The metric will be 1 even if
the Store succeeds, as long as at least one backend failed.