Separate DAS keyset and batch fetching logic
The logic for fetching keysets and batches was originally all done using
GetByHash on the the DASReader interface. Nitro nodes, batch posters,
and daservers would all include the ChainFetchReader at the beginning
of their chain of DASReaders in order to intercept GetByHash calls and
do the lookup of the keyset on L1. It was confusing, however, because it
also forwarded GetByHash requests to an inner DASReader, since it had no
way of telling if the GetByHash requests were for keysets or batch data.

This commit renames the ChainFetchReader to KeysetFetcher and removes the
inner DASReader and chaining logic so that it only tries to fetch keysets
from L1 or its own internal cache. The orchestration of normal nodes and
batch posters is changed to set up the KeysetFetcher separately rather
than part of the chain of DASReaders.

The replay binary preimage reader still needs to be updated.