Increase the GasLimit to 32000000.
I have no real understanding of how this fixes the problem. But, I
noticed that it was used a few lines above, and Tsahi also recommended
that increasing the limit could fix the tests.

This fixes the failing `TestProgramCacheManager` on my local MacOS
environment, but does not fix the `TestProgramMemory` which is failing
with a slightly different error.