Switch the dirty_indices to dirty_leaves
This provides a 32x reduction in the size of the dirty tracking.

It makes sense because if any index in the memory which contributes to
a leaf changes, then they hash for that leaf has to be
recalculated. So, we don't need to track the dirty indices in memory,
but the dirty leaves in the corresponding merkle tree.

This commit also removes a useless take() call, and documents some
invariantes when storing to memory.