Configurable timeout for HeaderReader polling
Previously if the backend stopped responding the
client.HeaderByNumber call inside the timed main loop would never
return, meaning that if the HeaderReader was only operating in polling
mode, calls to LastHeader would always return the last successfully
fetched header until OldHeaderTimeout, and if that was set to a high
duration then it would effectively always return the last successfully
fetched header and never an error. This would lead to the chain appear
to be stuck and not advancing to clients of HeaderReader.LastHeader.

This change adds a configurable timeout on this call, defaulting to 5s.
Other places in HeaderReader that use LastHeader use the passed-in
context which the client can control, so no extra timeout was added for
these.