Check all "not found" errors
Currently only the leveldb not found errors are handled. If a pebble not
found error is returned by the backend it is propagated, causing the
batchposter to fail to send transactions.

With this change all DB not found errors are handled the same way.

(cherry picked from commit 85068866fb3f409cebfc0a5c446bdab05e655d03)