Fix the deserialization of the UserWasm data
This change also:

1. Documents the `parse_inputs.rs` file in an attempt to give
context to why the deserialization is special.

2. Hardcodes the architecture to `wasm` when recording
ValidationInputs for use with the prover. Previously, the `jit`
validator would use the go system architecture, and the `arbitrator`
validator would use `wavm`.

3. Switches the `TestStorageProgram` test back to the `jit` validator.

Note: As of this commit, there is still a small problem that when
running the validator using the json inputs, it does not come up with
the same end has as `entry.End.Hash`.