Explode if more flags are specified with --dev
The --dev command-line flag silently expands to a number of other
command-line flags dropping any other flags that are passed with it.
This behavior can be very surprising to developers who are trying to
arrange a configuration from the command-line which is slightly
different from the flags to which --dev expands.

```
$> go run ./cmd/nitro --dev --pprof
```

This would run as though "--pprof" wasn't even set.

```
$> go run ./cmd/nitro --dev --pprof
```

The command exits with a configuration error explaining that --dev is
only valid when it is the only command-line flag.