Add a hacky sleep
Reviewers, please tell me the "right" way to ensure that the
transaction is visible in the L1. I don't like the idea of adding
sleeps in tests to ensure that some event has occurred. Is there some
RPC I can call or event I can subscribe to that I can then wait for
before proceeding?