Improve out of gas warning log
For the setTrieSlots and setTransientBytes32 HostIOs, when there is an
out-of-gas error, Nitro emits a warning line with `Caused by:\n \x02`.
This commit improves this error message by printing the debug
representation of the API status code and the warning becomes:
`Caused by:\n OutOfGas`.
This commit also solves a potential panic when the API does not return
the status code.