Add Mac OSX environment variables
The unfortunate thing is that it would still be good for developers to
set these in their shell environment so that if they do end up running
any "go" commands directly from the CLI, they won't be bombarded with
all the linker warnings. But, at least this makes the default
experience of using the make files a little nicer.