Disallow buggy stylus module imports
There are two different problems being addressed here.

1. `arbitrator_forwarder__` could be prepended to imports in user
stylus modules, and so long as the rest of the module import name was
a registered module, the user module coould be activated. But,
should the contract be called, the wrong opcode (`CrossModuleForwad`)
would be called which has problematic side-effects.

2. Any characters in `[/-\.]` could be substituted in a module import
where a registered module would have an underbar and module activation
would succeed. But, block execution of any calls the user stylus
module would fail because the import does not exist.

Examples:

* `import "vm_hooks" "arbitrator_forward__evm_gas_left"` would become
`import "vm_hooks" "evm_gas_left"`

* `import "vm/hooks" "evm_gas_left"` would become
`import vm_hooks" "evm_gas_left"`

The code which caused the first problem has simply been guarded by a
check that `allow_hostapi` is true. That will be false when
handling any stylus user modules.

The code which causes the second problem has simply been removed as
it is more ergonomical to simply allow modules with these special
characters in their names, than to replace them with `_`.