Update espresso-sequencer-go with new LC ABI (#230)
* Update espresso-sequencer-go with new LC ABI

Instead of using the block height from the first snapshot after the
height in the justification it uses the latest height in the LC
contract. Currently not 100% convinced about correctness but also
believe it should work because any newer tree still contains the header
of interest.

* Undo LC ABI changes

We're not comfortable with the previous change because making a proof
w.r.t to the newest finalized state may cause issues if the batch poster
is lagging behind. This is currently not easy to test so we decided to
revert back to the old flow.

* Use tag of espresso-sequencer-go