Fix timeboost auction resolution queue handling
If there were auction resolution transactions in the
timeboostAuctionResolutionTxQueue but no normal transactions came
through the txQueue, then the auction resolution tx would not be
processed because the main createBlock loop was waiting on the txQueue channel
and not the auction resolution queue. We had tried to use the same pattern
as we had used for the txRetryQueue but this is not correct as the
auction resolution queue, like the txQueue, can have items added to it
asynchronously, whereas the txRetryQueue is only added to from the
createBlock itself.

This commit makes the timeboostAuctionResolutionTxQueue also a channel
so that it can be waited on in the same select statement and handle
asynchronous events correctly.

This also fixes two issues related to shutdown. The first was a race condition
with shutting down and txRetryQueue handling, and the second was that
we were missing adding forwarding for outstanding auction resolution txs upon
shutdown.

Fixes NIT-2878