Attempt at fixing the virtual leaves issue
The CollectProof and CollectMachineHashes functions were both susceptible to
challenges where it was possible that the rival would have committed to more
messages than this validator. And, in that case, it would attempt to look up a
message number which was greater than the highest messge number it had verified
as part of the batch in which the challenge originated.