Handle challenges in the virtually padded part of the leaaves
Before this change, if there was a block challenge at a height in any block
above the batchLimit, then the validator was not correctly creating inclusion
proofs because it was attempting to fetch execution results for blocks which
didn't really exist.

Now, the code detects that situation and simply returns the hash of an
arbitrator machine in the FINISHED state (since the Virtual leaf hashes) are all
in that state by virtue of their being repeated copies of the end state of a
block.