Fix the CollectProof and CollectMachineHashes calls
Previously, they couldn't produce correct inclusion proofs for block-level
challenges on blocks in the virtual range. That is to say, if this validaotor
processed n L2Blocks when creating an assertion and a rival processed >= n+1
L2Blocks, this validator would attempt to lookup a block index for which no
real block existed.

Now, the code properly catches this case and the last machine state for the last
real block is used for all virtual L2Blocks.