Add FFI bindings for Espresso validation (#294)
* Example call to espresso-crypto function stub in batch poster.

This commit includes a log line and example call to the espresso-crypto
function stubs in batch_poster.go

* Result of `make fmt`

Run `make fmt` to clean up formatting not apparent in editor

* Make current lib.rs c abi compatible

* result of make fmt

* Add C ABI ffi between rust crypto lib and espressocrypto go module.

This commit adds FFI support between exsisting Rust crypto helper code
and the exsisting espressocrypto go module in native.go These functions
will now serve as utility functions to use in the integration that will
handle ffi to the Rust code.

This commit also moves the espresso-crypto-helper code to the
espressocrypto directory as we no longer need it for the arbitrator.

* Build native.go successfully

* Update the espresso-crypto and add test for merkle proof verification

* update flake.nix to new go version and fix errors

* Current state of work

The current state of my work revolves around editing the Makefile to
include the new espressocrypto ffi functions in the build process.
Additionally, this commit restores some espressocrypto code to the
arbitrator in an attempt to get a docker image built via the test node.
I felt as though this would be easier for this PR as opposed to ripping
out all of the code from the arbitrator.

* Update the flake.lock

* Add the namespace proof test

* Current state of build system changes

This commit represents the current state of the build system after
testing today.
Currently things compile locally, but not when building an image from
the docker file. There are errors related to duplicate function
definitions from what seem to be the FFI lib that stylus exposes.

* Build nitro docker image sucessfully

- Modify makefile to have a file target. Since the build artifact of our
  static library is a file we should use this as the make target
  (instead of only having PHONY targets).

- Use rust 1.81 to build all shared rust libraries in docker.

Our crate requires a minimum version of 1.81. However the other
libraries like libstylus.a are currently compiled with 1.75 and this
seems to cause issues with duplicate symbol definitions during linking.

For example:

    8.718 go build -ldflags="-extldflags '-ldl' -X github.com/offchainlabs/nitro/cmd/util/confighelpers.version= -X github.com/offchainlabs/nitro/cmd/util/confighelpers.datetime= -X github.com/offchainlabs/nitro/cmd/util/confighelpers.modified=" -o target/bin/nitro "/workspace/cmd/nitro"
    42.74 # github.com/offchainlabs/nitro/cmd/nitro
    42.74 /usr/local/go/pkg/tool/linux_amd64/link: running clang failed: exit status 1
    42.74 /usr/bin/ld: /workspace/espressocrypto/lib/espresso-crypto-helper/target/release/libespresso_crypto_helper.a(std-1c4b19562077c20d.std.1f2242ed6435445e-cgu.0.rcgu.o): in function `__rdl_alloc':
    42.74 /rustc/eeb90cda1969383f56a2637cbd3037bdf598841c/library/std/src/alloc.rs:397: multiple definition of `__rdl_alloc'; /workspace/execution/gethexec/../../target/lib/libstylus.a(std-90f6ddbf82de36ec.std.409886f6357001f0-cgu.0.rcgu.o):/rustc/82e1608dfa6e0b5569232559e3d385fea5a93112/library/std/src/alloc.rs:389: first defined here

Bump the rust version of the prover-header-builder to 1.81 and compile
the espresso crypto library in the same container.

* Fix LDFLAGS, remove unnecessary CFLAGS

* Fix some merge errors

- We need rust 1.81 to compile the espresso crypto lib
- Re-add makefile rules

* Fix merge errors

* Remove FFI call in batch poster to expidite merge.

* Update bf submodule in arbitrator to match upstream version (This was set to a different commit during my erroneous merge)

* Remove extra change in espressocrypto_wasm.go

* Fix error with temporary espressocryptowasm directory preventing CI from running

* fix error of missing function definitions in espressocryptowasm directory to allow CI to run

* Re-add nosec comment about gosec error that will be removed soon

* result of make fmt

---------

Co-authored-by: Zach Showalter <zacshowa@gmail.com>
Co-authored-by: ImJeremyHe <yiliang.he@qq.com>