Create ADR-0001 Avoid primative constraint types
This PR creates the ADR (Architectural Decision Record) structure in the docs in
our repository and adds the first two ADRs. The first is sort of a meta-ADR in
that it just captures our decision to actually use ADRs to record our
architecturally-significant decisions for posterity and which format to use.