Refactor RoundTimingInfo, fix auctionCloseTicker
The logic for the calculations of "when does the next round start?", "what
round is it?", "is the auction closed?", was spread throughout the code
and required passing around multiple separate variables that all come
from the RoundTimingInfo from the ExpressLaneAuction contract and should
never change. This commit adds a domain specific RoundTimingInfo that
validates these fields on construction and consolidates the
aforementioned calculations.

This commit also contanis a fix for auctionCloseTicker, now roundTicker.
It used to assume the initial Offset timestamp fetched from the
ExpressLaneAuction contract would always start on a minute boundary, and
the round length would be a minute which is not always the case. This
commit changes it to be called roundTicker and it now uses standard
methods on the new RoundTimingInfo which take into account all
possibilities for initial offset and round duration.