feat: disable sequencer MaxTxDataSize check against sequencer inbox MaxDataSize when DA is enabled
currently when DA is enabled the batch poster MaxSize check against the sequencer inbox MaxDataSize is disabled. hoping for the same behavior on the sequencer MaxTxDataSize check so that we can increase an orbit chain's max transaction size when Anytrust is enabled.

keeping the check of the sequencer MaxTxDataSize against the batchPoster MaxSize as i'm assuming this is needed even when using an alt DA.

by doing this, we can theoretically increase the sequencer MaxTxDataSize to ~212KB(limited by another check in the sequencer logic against arbostypes.MaxL2MessageSize) provided that we also bump the batchPoster MaxSize appropriately.

originally was going to create a new dangerous flag for the sequencer that would disable all of these checks. however, i think this is a potentially simpler fix as comparing the sequencer MaxTxDataSize with the sequencer inbox MaxDataSize is probably not needed when the Anytrust committment in the sequencer inbox is a small/fixed size. would greatly appreciate any details if this is not a correct assumption to make.